requirements.txt
pyspark==3.4.0
kafka-python==2.0.2
cassandra-driver==3.25.0
tweepy==4.10.1
textblob==0.17.1
pytest==7.3.1